pyproject.toml
[project]
name = "outpeer-students-course-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
   "asgiref==3.8.1",
   "django==5.1.6",
   "djangorestframework==3.15.2",
   "psycopg==3.2.5",
   "sqlparse==0.5.3",
]

